requirements.txt
Django==1.5.2
Pillow==2.1.0
South==0.8.2
argparse==1.2.1
braintree==2.23.0
django-autoslug==1.7.1
django-ckeditor==4.0.2
django-easy-maps==0.8.4
django-grappelli==2.4.6
django-pdf==0.1
django-registration==1.0
django-transmeta==0.7.1
geopy==0.95.1
html5lib==0.95
mock==1.0.1
pisa==3.0.33
pyPdf==1.13
reportlab==2.6
requests==1.2.3
six==1.3.0
wsgiref==0.1.2
xhtml2pdf==0.0.5
